COP4520 P1p1: Minor correction

+++ UCF/COP4520/ProgrammingAssignment1/problem1.c
@@ -33,7 +33,7 @@ void printTime();
 
 struct fiveBytes {
 	volatile uint32_t curValue;
-	volatile uint32_t isWaiting;
+	volatile uint8_t isWaiting;
 };
 
 int main()
